Spotify Authorization Working and Merged Into Main (#5)
Up to this point, I have learned a ton.  I taught myself how to use TypeScript, Vite, react-router-dom v6.24.0,  Redux.  This is in addition to reacquainting myself in-depth with Express.js, dotenv, cors, and React.  It's kind of frustrating how many tools and APIs you need to familiarize yourself with in modern web development, but I am proud of how far this project has taken me.